fix: make sure we honour config files
Because we're installing in, e.g., /opt/package/wget-1.21.4, the
config file is /opt/wget/wget-1.21.4/etc/wgetrc and not
/opt/etc/wgetrc. So, we need to make sure changes in /opt/etc
performed by the user are reflected back into the right dir
but we also need to make sure that upgrading a package doesn't
remove previous user-edited configurations. To this end, we
(1) transform all config files in /opt/package/wget-1.21.4 into
.new files and then (2) create /opt/etc/wgetrc from a new file
if it's missing and then (3) symlink from /opt/etc/wgetrc to
/opt/package/wget-1.21.4/etc/wgetrc. The end result is that the
user modifies the file they expect and wget sees the config
file in the right place. Apply these changes to p11-kit as well.